+++ odtp/dashboard/pages/page_user.py
@@ -1,11 +1,15 @@
-from nicegui import ui
+from nicegui import ui, app
+import json
+import logging
 
 import odtp.dashboard.utils.storage as storage
 import odtp.dashboard.utils.ui_theme as ui_theme
 import odtp.mongodb.db as db
+from odtp.dashboard.utils.file_picker import local_file_picker
+from odtp.helpers.settings import ODTP_PATH
 
 
-def content() -> None:
+def content() -> None:  
     ui.markdown(
         """
         # Manage Users 
@@ -41,9 +45,9 @@ def ui_users_select() -> None:
         user_options = {str(user["_id"]): user["displayName"] for user in users}
         current_user = storage.get_active_object_from_storage((storage.CURRENT_USER))
         if current_user:
-            value = current_user["user_id"]
+            value = str(current_user["user_id"])
         else:
-            value = ui_theme.NO_SELECTION_INPUT      
+            value = ""      
         ui.select(
             user_options,
             value=value,
@@ -52,9 +56,8 @@ def ui_users_select() -> None:
             with_input=True,
         ).props("size=80")
     except Exception as e:
-        ui.notify(
-            f"User selection could not be loaded. An Exception occured: {e}",
-            type="negative",
+        logging.error(
+            f"User selection could not be loaded. An Exception occured: {e}"
         )
 
 
@@ -105,6 +108,9 @@ def ui_workarea():
     )
     try:
         user = storage.get_active_object_from_storage(storage.CURRENT_USER)
+        workdir = storage.get_value_from_storage_for_key(storage.CURRENT_USER_WORKDIR)
+        if not workdir:
+            workdir = ODTP_PATH
         if not user:
             ui.markdown(
                 f"""
@@ -118,6 +124,7 @@ def ui_workarea():
             f"""
             #### Current Selection
             - **user**: {user.get("display_name")}
+            - **work directory**: {workdir}
 
             ##### Actions
 
@@ -130,25 +137,39 @@ def ui_workarea():
             on_click=lambda: ui.open(ui_theme.PATH_DIGITAL_TWINS),
             icon="link",
         )
+        ui.button(
+            "Set Work directory", 
+            on_click=pick_workdir, 
+            icon="folder"
+        )
+        ui.button(
+            "Reset Work directory to default", 
+            on_click=reset_workdir, 
+            icon="folder"
+        ).props('flat')       
     except Exception as e:
-        ui.notify(
-            f"Workarea could not be retrieved. An Exception occured: {e}",
-            type="negative",
+        logging.error(
+            f"Workarea could not be retrieved. An Exception occured: {e}"
         )
 
 
-def store_selected_user(value):
-    if value == ui_theme.NO_SELECTION_VALUE:
-        return
+def store_selected_user(user_id):
+    if user_id == "":
+        return 
     try:
-        storage.storage_update_user(user_id=value)
-        storage.reset_storage_keep([storage.CURRENT_USER])
+        user = db.get_document_by_id(
+            document_id=user_id, collection=db.collection_users
+        )
+        current_user = json.dumps(
+            {"user_id": user_id, "display_name": user.get("displayName")}
+        )
+        app.storage.user[storage.CURRENT_USER] = current_user        
     except Exception as e:
-        ui.notify(
-            f"Selected user could not be stored. An Exception occured: {e}",
-            type="negative",
+        logging.error(
+            f"Selected user could not be stored. An Exception happened: {e}"
         )
     else:
+        storage.reset_storage_keep([storage.CURRENT_USER])
         ui_workarea.refresh()
 
 
@@ -165,7 +186,7 @@ def add_user(name_input, github_input, email_input):
         ui.notify(f"A user with id {user_id} has been created", type="positive")
     except Exception as e:
         ui.notify(
-            f"The user could not be added in the database. An Exception occured: {e}",
+            f"The user could not be added in the database. An Exception occurred: {e}",
             type="negative",
         )
     else:
@@ -173,4 +194,23 @@ def add_user(name_input, github_input, email_input):
         ui_add_user.refresh()
 
 
+async def pick_workdir() -> None:
+    try:
+        root = ODTP_PATH
+        result = await local_file_picker(root, multiple=False)
+        if result:
+            workdir = result[0]
+            app.storage.user[storage.CURRENT_USER_WORKDIR] = workdir
+            ui.notify(f"A new user workdir has been set to {workdir}", type="positive")
+    except Exception as e:
+        logging.error(
+            f"Work directory could not be picked: an Exception occurred: {e}"
+        )
+    else:            
+        ui_workarea.refresh()    
 
+
+def reset_workdir():
+    app.storage.user[storage.CURRENT_USER_WORKDIR] = ODTP_PATH
+    ui.notify(f"User workdir has been set to {ODTP_PATH}", type="positive")
+    ui_workarea.refresh()