+++ odtp/dashboard/utils/helpers.py
@@ -1,13 +1,18 @@
-def get_workflow_mairmaid(step_names):
-    if len(step_names) == 1:
-        return step_names[0]
-    step_name_tuples = [
-        (step_names[i - 1], step_names[i]) for i in range(len(step_names)) if i > 0
-    ]
-    workflow_in_mermaid = ""
-    for step_tuple in step_name_tuples:
-        workflow_in_mermaid += f"{step_tuple[0]} --> {step_tuple[1]};"
-    return workflow_in_mermaid
+import odtp.mongodb.db as db
+
+
+def get_workflow_mermaid(step_names, init="graph TB;"):
+    mermaid_graph = init
+    step_count = len(step_names)
+    if step_count == 1:
+        mermaid_graph += f"{step_names[0]};"
+    elif step_count > 1:    
+        step_name_tuples = [
+            (step_names[i - 1], step_names[i]) for i in range(step_count) if i > 0
+        ]
+        for i, step_tuple in enumerate(step_name_tuples):
+            mermaid_graph += f"C{i}[{step_tuple[0]}] --> C{i+1}[{step_tuple[1]}];"
+    return mermaid_graph
 
 
 def pd_lists_to_counts(items_list):
@@ -20,9 +25,49 @@ def component_version_for_table(version):
     component = version.get("component")
     version_cleaned = {
         "component": component.get("componentName"),
+        "version": version.get("component_version"),        
         "repository": component.get("repoLink"),
+        "commit": version.get("commitHash")[:8],
         "type": component.get("type"),
-        "version": version.get("component_version"),
-        "commit": version.get("commitHash")[:8]
     }
     return version_cleaned
+
+
+def get_execution_step_display_name(
+    component_name, 
+    component_version, 
+):
+    display_name = f"{component_name}:{component_version}"
+    return display_name    
+
+
+def get_key_from_parameters(current_component_parameters, index):
+    if index >= len(current_component_parameters):
+        return ""
+    parameters_keys_as_list = list(current_component_parameters.keys())  
+    return parameters_keys_as_list[index]
+
+
+def get_value_from_parameters(current_component_parameters, index):
+    if index >= len(current_component_parameters):
+        return ""    
+    parameters_values_as_list = list(current_component_parameters.values())
+    return parameters_values_as_list[index]           
+
+
+def get_execution_select_options(digital_twin_id):
+    executions = db.get_sub_collection_items(
+        collection=db.collection_digital_twins,
+        sub_collection=db.collection_executions,
+        item_id=digital_twin_id,
+        ref_name=db.collection_executions,
+        sort_by=[("start_timestamp", db.DESCENDING)]
+    )        
+    if not executions:
+        return {}
+    execution_options = {}
+    for execution in executions:
+        execution_options[
+            str(execution["_id"])
+        ] = f"{execution['start_timestamp'].strftime('%d/%m/%y')} {execution.get('title')}"
+    return execution_options