+++ odtp/dashboard/utils/storage.py
@@ -1,7 +1,5 @@
 import json
-import os
-import re
-from datetime import datetime
+import logging
 
 from nicegui import app, ui
 
@@ -21,6 +19,11 @@ CURRENT_COMPONENT = "component"
 CURRENT_LOCAL_SETTINGS = "local_settings"
 FORM_STATE_START = "start"
 FORM_STATE_STEP = "step"
+CURRENT_USER_WORKDIR = "user_workdir"
+CURRENT_PROJECT_PATH = "project_path"
+EXECUTION_FOR_RUN = "execution_for_run"
+RUN_STEP = "run_step"
+SECRETS_FILES = "secrets_files"
 
 
 class ODTPFormValidationException(Exception):
@@ -36,13 +39,13 @@ def reset_storage_delete(keys):
             if key in current_storage_keys:
                 del app.storage.user[key]
     except Exception as e:
-        ui.notify(
+        logging.error(
             f"""During reset storage delete with keys {keys} en execption {e} occured. 
                   Current keys in storage {app.storage.user.keys()}"""
         )
 
 
-def reset_storage_keep(keys):
+def reset_storage_keep(keys):   
     current_storage_keys = app.storage.user.keys()
     try:
         if not isinstance(keys, list):
@@ -52,131 +55,10 @@ def reset_storage_keep(keys):
             if (key not in keys) and (key in current_storage_keys):
                 del app.storage.user[key]
     except Exception as e:
-        ui.notify(
+        logging.error(
             f"""During reset storage keep with keys {keys} en execption {e} occured. 
                   Current keys in storage {app.storage.user.keys()}"""
         )
-
-
-def storage_update_digital_twin(digital_twin_id):
-    try:
-        digital_twin = db.get_document_by_id(
-            document_id=digital_twin_id, collection=db.collection_digital_twins
-        )
-        current_digital_twin = json.dumps(
-            {"digital_twin_id": digital_twin_id, "name": digital_twin.get("name")}
-        )
-        app.storage.user[CURRENT_DIGITAL_TWIN] = current_digital_twin
-    except Exception as e:
-        ui.notify(
-            f"storage update for {CURRENT_DIGITAL_TWIN} failed: {e}", type="negative"
-        )
-
-
-def storage_update_local_settings(project_path):
-    try:
-        if CURRENT_LOCAL_SETTINGS in app.storage.user.keys():
-            local_settings = get_active_object_from_storage(CURRENT_LOCAL_SETTINGS)
-        else:
-            local_settings = {}
-        if project_path:
-            local_settings["project_path"] = project_path
-            app.storage.user[CURRENT_LOCAL_SETTINGS] = json.dumps(local_settings)
-    except Exception as e:
-        ui.notify(f"storage update for local settings failed: {e}", type="negative")
-
-
-def storage_update_execution(execution_id):
-    try:
-        execution = db.get_document_by_id(
-            document_id=execution_id, collection=db.collection_executions
-        )
-        version_names = odtp_utils.get_version_names_for_execution(execution)
-        current_execution = {
-            "execution_id": execution_id,
-            "title": execution.get("title"),
-            "timestamp": execution.get("start_timestamp").strftime(
-                "%m/%d/%Y, %H:%M:%S"
-            ),
-            "versions": execution.get("component_versions"),
-            "step_names": version_names,
-            "steps": [str(step_id) for step_id in execution.get("steps")],
-        }
-        current_execution_as_json = json.dumps(current_execution)
-        app.storage.user[CURRENT_EXECUTION] = current_execution_as_json
-    except Exception as e:
-        ui.notify(f"storage update for execution failed: {e}", type="negative")
-
-
-def storage_update_add_execution_init(
-    name, version_tuples, digital_twin_id, form_state, parameter_count
-):
-    if form_state == FORM_STATE_START:
-        version_ids = [version_tuple[0] for version_tuple in version_tuples]
-        step_names = [version_tuple[2] for version_tuple in version_tuples]
-        add_execution = {
-            "name": name,
-            "versions": version_ids,
-            "step_names": step_names,
-            "digital_twin_id": digital_twin_id,
-            "step_count": len(version_ids),
-            "parameter_count": str(parameter_count),
-            "current_step_nr": "0",
-            "parameters": [],
-            "ports": [],
-        }
-        app.storage.user[NEW_EXECUTION] = json.dumps(add_execution)
-
-
-def storage_update_add_execution_step(
-    parameter_keys,
-    parameter_values,
-    ports_mapping_inputs,
-):
-
-    new_execution = get_active_object_from_storage(NEW_EXECUTION)
-    if not len(parameter_keys) == len(parameter_values):
-        raise ODTPFormValidationException("paramter key and value count must be equal")
-    parameters_step = {}
-    ports_step = []
-    for key, value in zip(parameter_keys, parameter_values):
-        parameters_step[key] = value
-    ports_step = parse.parse_ports(ports_mapping_inputs)
-    new_execution["parameters"].append(parameters_step)
-    new_execution["ports"].append(ports_step)
-    current_step_nr = int(new_execution["current_step_nr"])
-    current_step_nr += 1
-    new_execution["current_step_nr"] = str(current_step_nr)
-    app.storage.user[NEW_EXECUTION] = json.dumps(new_execution)
-
-
-def storage_update_add_component(repo_link):
-    """the repo link that gets stored for the component is taken from the
-    github api, so that repos are not stored double"""
-    latest_commit = odtp_git.check_commit_for_repo(repo_link)
-    repo_info = odtp_git.get_github_repo_info(repo_link)
-    add_component = {
-        "repo_link": repo_info.get("html_url"),
-        "latest_commit": latest_commit,
-        "repo_info": repo_info,
-    }
-    try:
-        app.storage.user[NEW_COMPONENT] = json.dumps(add_component)
-    except Exception as e:
-        ui.notify(f"storage update for new component failed: {e}", type="negative")
-
-
-def storage_update_user(user_id):
-    try:
-        user = db.get_document_by_id(
-            document_id=user_id, collection=db.collection_users
-        )
-        current_user = json.dumps(
-            {"user_id": user_id, "display_name": user.get("displayName")}
-        )
-        app.storage.user[CURRENT_USER] = current_user
-    except Exception as e:
-        ui.notify(f"{CURRENT_USER} could not be set. Exception {e} occured")
 
 
 def get_active_object_from_storage(object_name):
@@ -185,10 +67,13 @@ def get_active_object_from_storage(object_name):
         if object:
             return json.loads(object)
     except Exception as e:
-        ui.notify(
-            f"'{object_name}' could not be retrieved from storage. Exception occured: {e}",
-            type="negative",
+        logging.error(
+            f"'{object_name}' could not be retrieved from storage. Exception occured: {e}"
         )
+
+
+def get_value_from_storage_for_key(storage_key):
+    return app.storage.user.get(storage_key)     
 
 
 def storage_update_component(component_id):
@@ -211,20 +96,48 @@ def storage_update_component(component_id):
         )
         app.storage.user[CURRENT_COMPONENT] = current_component
     except Exception as e:
-        ui.notify(
+        logging.error(
             f"storage update for {CURRENT_COMPONENT} failed: {e}", type="negative"
         )
 
 
-def storage_run_selection(execution_id, repo_url, commit_hash):
-    run_selection = {
+def store_execution_selection(storage_key, execution_id):  
+    execution = db.get_document_by_id(
+        document_id=execution_id, collection=db.collection_executions
+    )
+    version_tags = odtp_utils.get_version_names_for_execution(
+        execution=execution,
+        naming_function=helpers.get_execution_step_display_name,
+    )
+    step_ids = [str(step_id) for step_id in execution["steps"]]
+    step_documents = db.get_document_by_ids_in_collection(
+        document_ids=step_ids, collection=db.collection_steps
+    )
+    step_dict = {}
+    for step_document in step_documents:
+        step_dict[str(step_document["_id"])] = step_document
+    ports = []
+    parameters = []
+    outputs = []
+    inputs = []
+    for step_id in step_ids:
+        parameters.append(step_dict[step_id].get("parameters", {}))
+        ports.append(step_dict[step_id].get("ports", []))
+        outputs.append(step_dict[step_id].get("output", {}))
+        inputs.append(step_dict[step_id].get("input", {}))     
+    current_execution = {
         "execution_id": execution_id,
-        "repo_url": repo_url,
-        "commit_hash": commit_hash,
+        "title": execution.get("title"),
+        "timestamp": execution.get("start_timestamp").strftime(
+            "%m/%d/%Y, %H:%M:%S"
+        ),
+        "versions": execution.get("component_versions"),
+        "version_tags": version_tags,
+        "steps": step_ids,
+        "ports": ports,
+        "parameters": parameters,
+        "outputs": outputs,
+        "inputs": inputs,    
     }
-    if not run_selection:
-        app.storage.user["run_selection"] = "None"
-    try:
-        app.storage.user["run_selection"] = json.dumps(run_selection)
-    except Exception as e:
-        ui.notify(f"storage update for run selection failed: {e}", type="negative")
+    current_execution_as_json = json.dumps(current_execution)
+    app.storage.user[storage_key] = current_execution_as_json