+++ odtp/dashboard/utils/ui_theme.py
@@ -6,6 +6,8 @@ from nicegui import app, ui
 import odtp.dashboard.utils.helpers as helpers
 import odtp.dashboard.utils.storage as storage
 import odtp.helpers.utils as odtp_utils
+import odtp.mongodb.db as db
+
 
 PATH_ABOUT = "/"
 PATH_USERS = "/users"
@@ -20,8 +22,8 @@ NO_SELECTION_INPUT = None
 def menu() -> None:
     ui.link("About", PATH_ABOUT).classes(replace="text-white")
     ui.link("Users", PATH_USERS).classes(replace="text-white")
-    ui.link("Digital Twins", PATH_DIGITAL_TWINS).classes(replace="text-white")
     ui.link("Components", PATH_COMPONENTS).classes(replace="text-white")
+    ui.link("Digital Twins", PATH_DIGITAL_TWINS).classes(replace="text-white")
     ui.link("Executions", PATH_EXECUTIONS).classes(replace="text-white")
     ui.link("Run", PATH_RUN).classes(replace="text-white")
 
@@ -30,7 +32,7 @@ def menu() -> None:
 def frame(navtitle: str):
     """Custom page frame to share the same styling and behavior across all pages"""
     ui.colors(
-        primary="black", secondary="black", accent="black", positive="black"
+        primary="black", secondary="black", accent="black", positive="teal"
     )
     with ui.header().classes("justify-between text-white"):
         ui.label("OTDP").classes("font-bold")
@@ -39,10 +41,10 @@ def frame(navtitle: str):
     yield
 
 
-def ui_add_first(item_name, page_link):
-        ui.label(f"You need to select a {item_name} first").classes('text-lg')
+def ui_add_first(item_name, page_link, action="select"):
+        ui.label(f"You need to {action} {item_name} first").classes('text-lg')
         ui.button(
-            f"Select {item_name}",
+            f"{action} {item_name}",
             on_click=lambda: ui.open(page_link),
             icon="link",
         )
@@ -50,3 +52,33 @@ def ui_add_first(item_name, page_link):
 
 def ui_no_items_yet(item_name_plural):
     ui.label(f"There are no {item_name_plural} yet. Start adding {item_name_plural}").classes('text-lg')
+
+
+def ui_execution_display(
+    execution_title,
+    version_tags,
+    ports,
+    parameters,
+):
+    ui.label(f"Execution Title: {execution_title}").classes("text-lg w-full")
+    table_columns = [
+        {'name': 'key', 'label': 'Parameter type / name', 'field': 'key'},
+        {'name': 'value', 'label': 'value', 'field': 'value'},
+    ]    
+    if not version_tags:
+        return  
+    with ui.grid(columns=2):
+        with ui.column():    
+            ui.mermaid(
+                helpers.get_workflow_mermaid(version_tags, init='graph TB;')
+            )
+        with ui.column(): 
+            for k, version_tag in enumerate(version_tags):
+                rows = []
+                if ports and ports[k]:    
+                    for port_mapping in ports[k]:
+                        rows.append({'key': 'port-mapping', 'value': port_mapping})
+                if parameters and parameters:
+                    for key, value in parameters[k].items():
+                        rows.append({'key': key, 'value': value},)
+                ui.table(columns=table_columns, rows=rows, row_key='key', title=version_tag)